requirements.txt
supabase
opencv-python
numpy
python-dotenv